move LMP stuff to its own helper
